requirements.txt
-i https://pypi.org/simple
apns2==0.5.0
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3 
cryptography==2.8
defusedxml==0.6.0
django-allauth==0.39.1
django-cors-headers==3.0.2
django-filter==2.2.0
django-push-notifications==1.6.1
django-rest-auth==0.9.5
django-taggit==1.2.0
django==2.2.6
djangorestframework-bulk==0.2.1
djangorestframework==3.9.4
fcm-django==0.3.4
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
lxml==4.5.0
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
psycopg2==2.8.4
py-vapid==1.7.0
pycodestyle==2.5.0
pycparser==2.20
pyfcm==1.4.7
pyjwt==1.7.1
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
pywebpush==1.10.1
requests-oauthlib==1.2.0
requests==2.22.0
six==1.12.0
soupsieve==2.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
